`nukedir` module implementation - extracted the logic from `__init__`, set up basic unit tests, no change in functionality yet.
