+++ Makefile
@@ -96,6 +96,7 @@ test: ALWAYS
 	@make local
 	pytest -v ./tests/poof-test.py
 	pytest -v ./tests/launchd-test.py
+	pytest -v ./tests/nukedir-test.py
 	pip uninstall -y $(PACKAGE)==$(VERSION) || true
 	rm -Rfv $$(find $(PACKAGE)/ | awk '/__pycache__$$/')
 	rm -Rfv $$(find tests | awk '/__pycache__$$/')
